Remove ScrollHelper by merging it with InfiniteList.
