+++ client/components/infinite-list/index.jsx
@@ -17,7 +17,6 @@ import afterLayoutFlush from 'lib/after-layout-flush';
  */
 import detectHistoryNavigation from 'lib/detect-history-navigation';
 import ScrollStore from './scroll-store';
-import ScrollHelper from './scroll-helper';
 import scrollTo from 'lib/scroll-to';
 import smartSetState from 'lib/react-smart-set-state';
 
@@ -45,7 +44,6 @@ export default class InfiniteList extends React.Component {
 
 	lastScrollTop = -1;
 	scrollRAFHandle = false;
-	scrollHelper = null;
 	isScrolling = false;
 	_isMounted = false;
 	smartSetState = smartSetState;
@@ -66,11 +64,10 @@ export default class InfiniteList extends React.Component {
 			newState.scrollTop = scrollTop;
 		}
 
-		this.scrollHelper = new ScrollHelper( this.boundsForRef );
-		this.scrollHelper.props = this.props;
+		this.initScrolling();
 		if ( this._contextLoaded() ) {
-			this._scrollContainer = this.scrollHelper.props.context || window;
-			this.scrollHelper.updateContextHeight( this.getCurrentContextHeight() );
+			this._scrollContainer = this.props.context || window;
+			this.updateContextHeight( this.getCurrentContextHeight() );
 		}
 
 		if ( newState ) {
@@ -80,7 +77,7 @@ export default class InfiniteList extends React.Component {
 			newState = {
 				firstRenderedIndex: 0,
 				topPlaceholderHeight: 0,
-				lastRenderedIndex: this.scrollHelper.initialLastRenderedIndex(),
+				lastRenderedIndex: this.initialLastRenderedIndex(),
 				bottomPlaceholderHeight: 0,
 				scrollTop: 0,
 			};
@@ -107,8 +104,6 @@ export default class InfiniteList extends React.Component {
 	}
 
 	componentDidUpdate( prevProps ) {
-		this.scrollHelper.props = this.props;
-
 		if ( ! this._contextLoaded() ) {
 			return;
 		}
@@ -141,11 +136,10 @@ export default class InfiniteList extends React.Component {
 	reset() {
 		this.cancelAnimationFrame();
 
-		this.scrollHelper = new ScrollHelper( this.boundsForRef );
-		this.scrollHelper.props = this.props;
+		this.initScrolling();
 		if ( this._contextLoaded() ) {
 			this._scrollContainer = this.props.context || window;
-			this.scrollHelper.updateContextHeight( this.getCurrentContextHeight() );
+			this.updateContextHeight( this.getCurrentContextHeight() );
 		}
 
 		this.isScrolling = false;
@@ -153,7 +147,7 @@ export default class InfiniteList extends React.Component {
 		this.setState( {
 			firstRenderedIndex: 0,
 			topPlaceholderHeight: 0,
-			lastRenderedIndex: this.scrollHelper.initialLastRenderedIndex(),
+			lastRenderedIndex: this.initialLastRenderedIndex(),
 			bottomPlaceholderHeight: 0,
 			scrollTop: 0,
 		} );
@@ -248,19 +242,19 @@ export default class InfiniteList extends React.Component {
 
 		this.lastScrollTop = this.getCurrentScrollTop();
 		ScrollStore.storeScrollTop( url, this.lastScrollTop );
-		this.scrollHelper.updateContextHeight( this.getCurrentContextHeight() );
-		this.scrollHelper.scrollPosition = this.lastScrollTop;
-		this.scrollHelper.triggeredByScroll = options.triggeredByScroll;
-		this.scrollHelper.updatePlaceholderDimensions();
+		this.updateContextHeight( this.getCurrentContextHeight() );
+		this.scrollPosition = this.lastScrollTop;
+		this.triggeredByScroll = options.triggeredByScroll;
+		this.updatePlaceholderDimensions();
 
-		this.scrollHelper.scrollChecks( this.state );
+		this.triggerScrollChanges( this.state );
 
-		if ( this.scrollHelper.stateUpdated ) {
+		if ( this.stateUpdated ) {
 			newState = {
-				firstRenderedIndex: this.scrollHelper.firstRenderedIndex,
-				topPlaceholderHeight: this.scrollHelper.topPlaceholderHeight,
-				lastRenderedIndex: this.scrollHelper.lastRenderedIndex,
-				bottomPlaceholderHeight: this.scrollHelper.bottomPlaceholderHeight,
+				firstRenderedIndex: this.firstRenderedIndex,
+				topPlaceholderHeight: this.topPlaceholderHeight,
+				lastRenderedIndex: this.lastRenderedIndex,
+				bottomPlaceholderHeight: this.bottomPlaceholderHeight,
 				scrollTop: this.lastScrollTop,
 			};
 
@@ -328,8 +322,6 @@ export default class InfiniteList extends React.Component {
 	}
 
 	render() {
-		this.scrollHelper.props = this.props;
-
 		const propsToTransfer = omit( this.props, Object.keys( this.constructor.propTypes ) ),
 			spacerClassName = 'infinite-list__spacer';
 		let i,
@@ -343,7 +335,7 @@ export default class InfiniteList extends React.Component {
 				this.props.items.length
 			);
 			lastRenderedIndex = Math.min(
-				this.state.firstRenderedIndex + this.scrollHelper.initialLastRenderedIndex(),
+				this.state.firstRenderedIndex + this.initialLastRenderedIndex(),
 				this.props.items.length - 1
 			);
 			debug( 'reset lastRenderedIndex to %s', lastRenderedIndex );
@@ -414,5 +406,464 @@ export default class InfiniteList extends React.Component {
 	 */
 	_contextLoaded() {
 		return this.props.context || this.props.context === false || ! ( 'context' in this.props );
+	}
+
+	/* SCROLL HELPER */
+
+	initScrolling() {
+		this.itemHeights = {};
+
+		// Hide levels and context height
+		this.contextHeight = null;
+		this.topHideLevelHard = null;
+		this.topHideLevelSoft = null;
+		this.bottomHideLevelHard = null;
+		this.bottomHideLevelSoft = null;
+		this.bottomHideLevelUltraSoft = null;
+
+		// set by component
+		this.scrollPosition = null;
+
+		// queried directly from placeholder rects
+		this.containerTop = null;
+		this.topPlaceholderHeight = null;
+		this.bottomPlaceholderHeight = null;
+		this.containerBottom = null;
+
+		this.stateUpdated = null;
+		this.firstRenderedIndex = null;
+		this.lastRenderedIndex = null;
+	}
+
+	storedItemHeight( itemKey ) {
+		let height = this.props.guessedItemHeight;
+
+		if ( itemKey in this.itemHeights ) {
+			height = this.itemHeights[ itemKey ];
+		}
+
+		return height;
+	}
+
+	forEachInRow( index, callback, context ) {
+		if ( typeof callback !== 'function' ) {
+			return;
+		}
+
+		if ( context ) {
+			callback = callback.bind( context );
+		}
+
+		const firstIndexInRow = index - ( index % this.props.itemsPerRow ),
+			lastIndexInRow =
+				Math.min( firstIndexInRow + this.props.itemsPerRow, this.props.items.length ) - 1;
+		for ( let i = firstIndexInRow; i <= lastIndexInRow; i++ ) {
+			callback( this.props.items[ i ], i );
+		}
+	}
+
+	storeRowItemHeights( fromDirection, index ) {
+		this.forEachInRow(
+			index,
+			function( item ) {
+				const itemKey = this.props.getItemRef( item ),
+					itemBounds = this.boundsForRef( itemKey );
+				let height;
+
+				if ( itemBounds ) {
+					if ( 'bottom' === fromDirection ) {
+						height = this.containerBottom - this.bottomPlaceholderHeight - itemBounds.top;
+					} else {
+						height = itemBounds.bottom - ( this.containerTop + this.topPlaceholderHeight );
+					}
+				} else {
+					height = this.props.guessedItemHeight;
+				}
+
+				this.itemHeights[ itemKey ] = height;
+			},
+			this
+		);
+	}
+
+	deleteRowItemHeights( index ) {
+		this.forEachInRow(
+			index,
+			item => {
+				const itemKey = this.props.getItemRef( item );
+				delete this.itemHeights[ itemKey ];
+			},
+			this
+		);
+	}
+
+	getRowHeight( index ) {
+		let maxHeight = 0;
+
+		this.forEachInRow(
+			index,
+			item => {
+				const itemKey = this.props.getItemRef( item ),
+					height = this.storedItemHeight( itemKey );
+
+				maxHeight = Math.max( maxHeight, height );
+			},
+			this
+		);
+
+		return maxHeight;
+	}
+
+	updateContextHeight( contextHeight ) {
+		if ( this.contextHeight === contextHeight ) {
+			return;
+		}
+
+		this.contextHeight = contextHeight;
+
+		this.topHideLevelHard = Math.min( -1 * contextHeight, -5 * this.props.guessedItemHeight );
+
+		this.topHideLevelSoft = Math.min( -2 * contextHeight, -10 * this.props.guessedItemHeight );
+
+		this.bottomHideLevelHard =
+			contextHeight + Math.max( contextHeight, 5 * this.props.guessedItemHeight );
+
+		this.bottomHideLevelSoft =
+			contextHeight + Math.max( 2 * contextHeight, 10 * this.props.guessedItemHeight );
+
+		this.bottomHideLevelUltraSoft =
+			contextHeight + Math.max( 3 * contextHeight, 15 * this.props.guessedItemHeight );
+	}
+
+	initialLastRenderedIndex() {
+		return Math.min(
+			this.props.items.length - 1,
+			Math.floor( this.bottomHideLevelSoft / this.props.guessedItemHeight ) - 1
+		);
+	}
+
+	updatePlaceholderDimensions() {
+		const topPlaceholderRect = this.boundsForRef( 'topPlaceholder' ),
+			bottomPlaceholderRect = this.boundsForRef( 'bottomPlaceholder' );
+
+		this.topPlaceholderHeight = topPlaceholderRect.height;
+		this.containerTop = topPlaceholderRect.top;
+
+		this.bottomPlaceholderHeight = bottomPlaceholderRect.height;
+		// It is important to use placeholder to get container bottom.
+		// Container node is reported longer than it should be in mobile Safari 7.0
+		this.containerBottom = bottomPlaceholderRect.bottom;
+	}
+
+	triggerScrollChanges( state ) {
+		this.reset( state );
+
+		this.adjustLastRenderedIndex();
+
+		if ( this.shouldHideItemsAbove() ) {
+			this.hideItemsAbove();
+		} else if ( this.shouldShowItemsAbove() ) {
+			this.showItemsAbove();
+		}
+
+		if ( this.shouldHideItemsBelow() ) {
+			this.hideItemsBelow();
+		} else if ( this.shouldShowItemsBelow() ) {
+			this.showItemsBelow();
+		}
+
+		if ( this.shouldLoadNextPage() ) {
+			this.loadNextPage();
+		}
+	}
+
+	reset( state ) {
+		this.stateUpdated = false;
+		this.firstRenderedIndex = state.firstRenderedIndex;
+		this.lastRenderedIndex = state.lastRenderedIndex;
+	}
+
+	adjustLastRenderedIndex() {
+		// last index -1 means everything is rendered - it can happen when
+		// item count is not known when component is mounted
+		const offset = this.initialLastRenderedIndex(),
+			lastIndex = this.lastRenderedIndex || 0, // fixes NaN
+			firstIndex = this.firstRenderedIndex || 0, // fixes NaN
+			itemCount = this.props.items.length;
+		let newIndex = lastIndex;
+
+		if ( itemCount === 0 ) {
+			newIndex = -1;
+		}
+
+		if ( lastIndex >= itemCount ) {
+			newIndex = Math.min( firstIndex + offset, itemCount - 1 );
+		}
+
+		if ( newIndex === -1 && itemCount > 0 && firstIndex === 0 ) {
+			newIndex = offset;
+		}
+
+		if ( newIndex !== this.lastRenderedIndex ) {
+			this.stateUpdated = true;
+			this.lastRenderedIndex = newIndex;
+		}
+	}
+
+	shouldHideItemsAbove() {
+		//
+		// Hiding Item Above Chart
+		//
+		//       +---- container top relative to context - value below zero
+		//       |
+		//       |  placeholder
+		//       |
+		//       +---- placeholder bottom edge (before) = container top + placeholder height
+		//       |
+		//       | item to be hidden
+		//       |
+		//       +----
+		//       |
+		//  -----|- soft hide limit = - 2 * context height
+		//       |
+		//       | item to be hidden
+		//       |
+		//       +----
+		//       |
+		//       | last item to be hidden
+		//       |
+		//       +---- new placeholder bottom edge
+		//       |
+		//  -----|- hard hide limit = -1 * context height
+		//       |
+		//       | this item will stay
+		//       |
+		//       +----
+		//       |
+		//       |
+		//       |
+		//       |
+		//  -----|- context top = 0
+		//       |
+		//
+		return this.containerTop + this.topPlaceholderHeight < this.topHideLevelSoft;
+	}
+
+	hideItemsAbove() {
+		let rowHeight, rowBottom;
+
+		while ( this.firstRenderedIndex < this.props.items.length ) {
+			this.storeRowItemHeights( 'top', this.firstRenderedIndex );
+			rowHeight = this.getRowHeight( this.firstRenderedIndex );
+			rowBottom = this.containerTop + this.topPlaceholderHeight + rowHeight;
+
+			if ( rowBottom > this.topHideLevelHard ) {
+				this.deleteRowItemHeights( this.firstRenderedIndex );
+				break;
+			}
+
+			this.topPlaceholderHeight += rowHeight;
+			this.firstRenderedIndex += this.props.itemsPerRow;
+			this.stateUpdated = true;
+			debug( 'hiding top item', rowHeight, this.topPlaceholderHeight );
+		}
+	}
+
+	shouldShowItemsAbove() {
+		//
+		// Showing Item Above Chart
+		//
+		//       +---- container top relative to context - value below zero
+		//       |
+		//       |
+		//       |
+		//       |
+		//  -----|- soft hide limit = - 2 * context height
+		//       |
+		//       +---- new placeholder bottom
+		//       |
+		//       | Last item to be shown
+		//       |
+		//       +----
+		//       |
+		//  -----|- hard hide limit = -1 * context height
+		//       |
+		//       +----
+		//       |
+		//       | Item to be shown
+		//       |
+		//       +---- placeholder bottom when check started
+		//       |
+		//  -----|- context top = 0
+		//       |
+		//
+		return this.containerTop + this.topPlaceholderHeight > this.topHideLevelHard;
+	}
+
+	showItemsAbove() {
+		let rowHeight, newPlaceholderBottom;
+
+		while ( this.firstRenderedIndex > 0 ) {
+			rowHeight = this.getRowHeight( this.firstRenderedIndex - this.props.itemsPerRow );
+			newPlaceholderBottom = this.containerTop + this.topPlaceholderHeight - rowHeight;
+
+			if ( newPlaceholderBottom < this.topHideLevelSoft ) {
+				break;
+			}
+
+			this.deleteRowItemHeights( this.firstRenderedIndex - this.props.itemsPerRow );
+			this.firstRenderedIndex -= this.props.itemsPerRow;
+			this.firstRenderedIndex = Math.max( 0, this.firstRenderedIndex );
+			if ( this.firstRenderedIndex <= 0 ) {
+				// never allow top placeholder when everything is shown
+				this.topPlaceholderHeight = 0;
+			} else {
+				this.topPlaceholderHeight = Math.max( 0, this.topPlaceholderHeight - rowHeight );
+			}
+
+			this.stateUpdated = true;
+			debug( 'showing top item', rowHeight, this.topPlaceholderHeight );
+		}
+	}
+
+	shouldHideItemsBelow() {
+		//
+		// Hiding Items Below Chart
+		//
+		//       |
+		//  -----|- context bottom = 1 * context height, e.g. 1000
+		//       |
+		//       +----
+		//       |
+		//       | Item
+		//       |
+		//       +----
+		//       |
+		//  -----|- hard hide limit, e.g. 2000
+		//       |
+		//       +---- new placeholder top
+		//       |
+		//       | Last item to be hidden
+		//       |
+		//       +----
+		//       |
+		//  -----|- soft hide limit, e.g. 3000
+		//       |
+		//       +----
+		//       |
+		//       | Item to be hidden
+		//       |
+		//       +----
+		//       |
+		//  -----|- 3rd hide limit, e.g. 4000
+		//       |
+		//       | Item to be hidden
+		//       |
+		//       +---- placeholder top when check started
+		//       |
+		//       |
+		//       | Bottom placeholder
+		//       |
+		//       +---- container bottom relative to context, e.g. 5000
+		//
+		const placeholderTop = this.containerBottom - this.bottomPlaceholderHeight;
+		return placeholderTop > this.bottomHideLevelUltraSoft;
+	}
+
+	hideItemsBelow() {
+		let rowTop, rowHeight;
+
+		while ( this.lastRenderedIndex >= 0 ) {
+			this.storeRowItemHeights( 'bottom', this.lastRenderedIndex );
+			rowHeight = this.getRowHeight( this.lastRenderedIndex );
+			rowTop = this.containerBottom - this.bottomPlaceholderHeight - rowHeight;
+
+			if ( rowTop < this.bottomHideLevelHard ) {
+				this.deleteRowItemHeights( this.lastRenderedIndex );
+				break;
+			}
+
+			this.bottomPlaceholderHeight += rowHeight;
+			this.lastRenderedIndex -= this.props.itemsPerRow;
+			this.stateUpdated = true;
+			debug( 'hiding bottom item', rowHeight, this.bottomPlaceholderHeight );
+		}
+	}
+
+	shouldShowItemsBelow() {
+		const placeholderTop = this.containerBottom - this.bottomPlaceholderHeight;
+		return placeholderTop < this.bottomHideLevelHard;
+	}
+
+	showItemsBelow() {
+		let rowHeight, itemTop, placeholderTop;
+
+		while ( this.lastRenderedIndex < this.props.items.length - 1 ) {
+			rowHeight = this.getRowHeight( this.lastRenderedIndex + this.props.itemsPerRow );
+			placeholderTop = this.containerBottom - this.bottomPlaceholderHeight;
+			itemTop = placeholderTop + rowHeight;
+
+			if (
+				itemTop > this.bottomHideLevelSoft &&
+				// always show at least one item when placholder top is above hard limit
+				placeholderTop > this.bottomHideLevelHard
+			) {
+				break;
+			}
+
+			this.deleteRowItemHeights( this.lastRenderedIndex + this.props.itemsPerRow );
+			if ( this.bottomPlaceholderHeight - rowHeight < 0 ) {
+				this.containerBottom += rowHeight - this.bottomPlaceholderHeight;
+				this.bottomPlaceholderHeight = 0;
+			} else {
+				this.bottomPlaceholderHeight -= rowHeight;
+			}
+			this.lastRenderedIndex += this.props.itemsPerRow;
+			this.lastRenderedIndex = Math.min( this.lastRenderedIndex, this.props.items.length - 1 );
+
+			// if everything is shown, then there should be no placeholder
+			if ( this.lastRenderedIndex >= this.props.items.length - 1 ) {
+				this.bottomPlaceholderHeight = 0;
+			}
+
+			this.stateUpdated = true;
+			debug( 'showing bottom item', rowHeight, this.bottomPlaceholderHeight );
+		}
+	}
+
+	shouldLoadNextPage() {
+		if ( this.props.fetchingNextPage || this.props.lastPage ) {
+			return false;
+		}
+
+		return this.bottomPlaceholderHeight === 0 && this.containerBottom < this.bottomHideLevelHard;
+	}
+
+	loadNextPage() {
+		if ( this.queuedFetchNextPage ) {
+			return;
+		}
+		let triggeredByScroll = this.triggeredByScroll;
+
+		debug( 'fetching next page', this.containerBottom, this.bottomPlaceholderHeight );
+
+		// Consider all page fetches once user starts scrolling as triggered by scroll
+		// Same condition check is in components/infinite-scroll checkScrollPosition
+		if ( this.scrollPosition > this.contextHeight ) {
+			triggeredByScroll = true;
+		}
+
+		// scroll check may be triggered while dispatching an action,
+		// we cannot create new action while dispatching old one
+		this.queuedFetchNextPage = Promise.resolve().then( () => {
+			this.queuedFetchNextPage = null;
+			// checking these values again because we shifted the fetch to the next stack
+			if ( this.props.fetchingNextPage || this.props.lastPage ) {
+				return false;
+			}
+			this.props.fetchNextPage( {
+				triggeredByScroll: triggeredByScroll,
+			} );
+		} );
 	}
 }
